Move to_table in database
