+++ src/database/database.rs
@@ -7,9 +7,10 @@ use xml::reader::XmlEvent::{CData, Characters, Comment, EndDocument, EndElement,
 
 use crate::database::decrypt::decrypt;
 use crate::database::field::Field;
-use crate::utils::password;
+use crate::utils::{format, password};
 
 use super::card::Card;
+use super::field;
 
 pub struct Database {
     cards: Vec<Card>
@@ -78,6 +79,47 @@ impl Database {
 
     pub fn get_cards(&self) -> &Vec<Card> {
         &self.cards
+    }
+
+    pub fn to_table(&self, passwords: bool) -> format::Table {
+        let mut table = Vec::new();
+
+        for card in self.get_cards() {
+            let mut row = Vec::new();
+
+            row.push((None, card.get_id().to_string()));
+            row.push((None, card.get_title().to_string()));
+
+            let mut i = 0;
+            for field in card.get_fields() {
+                i += 1;
+                if i > 3 {
+                    break;
+                }
+
+                let mask = "******";
+
+                let value = match field.get_type() {
+                    field::Type::Password | field::Type::Secret | field::Type::Pin => {
+                        if passwords {
+                            field.get_value()
+                        } else {
+                            Some("******")
+                        }
+                    }
+                    _ => field.get_value()
+                };
+                let value = value.unwrap_or("").to_string();
+
+                let name = field.get_name().to_string().clone();
+
+                row.push((Some(name), value));
+            }
+
+            table.push(row);
+        }
+
+        table
     }
 }
